Fix warnings, implement basic config system
