+++ fw/makefile
@@ -8,7 +8,7 @@ F_CPU = 16000000
 F_USB = $(F_CPU)
 OPTIMIZATION = s
 TARGET = beef
-SRC = $(TARGET).c Descriptors.c $(LUFA_SRC_USB) $(LUFA_SRC_USBCLASS) ws2812.c
+SRC = $(wildcard *.c) $(LUFA_SRC_USB) $(LUFA_SRC_USBCLASS)
 LUFA_PATH = LUFA
 CC_FLAGS = -DUSE_LUFA_CONFIG_HEADER -IConfig/
 LD_FLAGS =
