+++ fw/pin.c
@@ -1,0 +1,10 @@
+#include "beef.h"
+
+hw_pin hw_pins[] = {
+  { &DDRA, &PINA, &PORTA },
+  { &DDRB, &PINB, &PORTB },
+  { &DDRC, &PINC, &PORTC },
+  { &DDRD, &PIND, &PORTD },
+  { &DDRE, &PINE, &PORTE },
+  { &DDRF, &PINF, &PORTF },
+};
